Final commit

+++ src/ru/nsu/Filters/FilterNegative.java
@@ -12,9 +12,9 @@ public class FilterNegative extends Filter {
         for ( int i = 0; i<256; i++)
             for (int j = 0; j< 256; j++) {
                 new_array[j][i] = new MyRGB();
-                new_array[j][i].setRed((byte) (255 - array[j][i].getRed()));
-                new_array[j][i].setGreen((byte) (255 - array[j][i].getGreen()));
-                new_array[j][i].setBlue((byte) (255 - array[j][i].getBlue()));
+                new_array[j][i].setRed(255 - array[j][i].getRed());
+                new_array[j][i].setGreen(255 - array[j][i].getGreen());
+                new_array[j][i].setBlue(255 - array[j][i].getBlue());
             }
         return new_array;
     }
